Lots of progress.
This commit provides a new implementation of the FST node registry
that works much better on the DOI URL data set. In particular, it
can now cache any node instead of being limited to nodes with no more
than 2 transitions. This is done without any performance penalty and
a small constant factor increase in memory.

We also start (finally) writing code to do searching and kick it off
with a benchmark. Next up are range queries.