Add support for creating an Fst from a slice of Vec<u8>.
A new constructor method for Builder, called from_shared_bytes, has been
added in this commit. Its similar to from_bytes, but instead of taking
a Vec<u8>, it takes an Arc<Vec<u8>>, an offset and a length and creates an
Fst at that region of memory. This permits storing multiple Fsts inside
a single Vec<u8> in memory.