Fix output value bug.
When the output value is zero, we can't actually completely ignore it.
Consider, for example, these keys and values:

  a,1
  ab,0

When the second key is inserted, the common prefix between `1` and `0`
if `0`, which means `a`'s value of `1` needs to get moved to the node
specific "final output" field. If we don't do this, then `1` is treated
as if it were part of the summation of values corresponding to `ab`'s
value, which it isn't.

We fix this by making the internal API distinguish between `Output::zero()`
and "no output." We need to make sure that fst::Set therefore never
tries to insert an `Output::zero()`, since that would result in a
performance hit.

Fixes #35