split regexes and fuzzy search into separate crates
This splits the regular expression and fuzzy search support out into
separate crates. This was motivated by a desire to make the main fst
crate a bit smaller, particularly with its required dependencies. In
particular, it is desirable to potentially use it inside the
regex-syntax crate itself to represent Unicode tables.

With this change, it's possible to use the fst crate with only a
byteorder dependency by disabling default features.

This is a breaking change. Users of regular expressions and fuzzy
searching will need to change their code to use the fst-regex and/or
fst-levenshtein crates.