fst: core performance improvements
This commit makes two classes of performance improvements:

  1. Aggressively inline functions that read the FST's byte
     representation. The `inline(always)` annotations are a faux pas,
     but without them, the compiler appears to have a lot of difficulty
     inlining. Ideally, most of the `Fst::get` implementation should
     be inlined.
  2. Add a byte-lookup index for nodes with a lot of transitions.
     Previously, looking up a transition by a byte required linearly
     searching the transitions in each node. When a popular node has
     lots of transitions, this linear search becomes a pessimization
     since many paths go through the call to `get`, and each one might
     need to examine many transitions. The change in this commit is to
     add a 256 byte index to nodes with lots of transitions (32 or
     more), which lowers the constant on transition-lookup-by-byte such
     that it is not proportional to the alphabet size.

In order for optimization (2) to work, we needed to bump the FST format
version from 1 to 2. Backwards compatibility is preserved, but not
forwards compatibility. That means:

  fst<=0.2 cannot read FSTs generated by fst>=0.3.
  fst>=0.3 can read FSTs generated by fst<=0.2.

In the former case, an error will be returned if an older version of the
fst crate tries to read an FST generated by a newer version of the FST
crate. Therefore, even though forward compatibility is not maintained,
you'll get a hard failure instead of implicit data corruption.