raw: add as_bytes method
It's like to_vec, but without making copies.

The reason why it didn't already exist is because the &[u8] tends to be a
memory map, or at the very least, could be a memory map. In that circumstance,
getting access to the raw slice without needing to type unsafe would be bad.
With that said, since the last release, constructing an FST from a memory map
now requires typing unsafe. There is even a safe method called as_bytes on the
MmapReadOnly type. So to that end, the absence of this method was likely an
oversight, and should now be an okay addition.

PR #77