Add `fst::Str` automaton for exact strings
This adds ability to match exact strings against the given set, as well as enables trie usecase by combining it with `StartsWith<_>`.

Name is subject to change, I don't have strong preferences either way, as well as would be happy with just implementing `Automaton` directly on `str` / `[u8]` (which seems to make sense).