levenshtein/regex: move levenshtein, remove regex
This is where the Levenshtein automata originally lived. I originally
moved it out of the crate in order to reduce dependencies. But we can
instead just make it an optional feature.

Instead of moving fst-regex back into the crate as well, we instead
recommend using `regex-automata`. `regex-automata` is basically the
productionized version of "compile a regex to a DFA," which is exactly
what we want.