fst-bin: update dependencies and add some polish
Mostly this just upgrades to the newest versions of everything, and
replaces deprecated crates (tempdir, chan) with their newer variants
(tempfile, crossbeam-channel).

We also use BString in several places instead of String so as to permit
indexing any keys and not just UTF-8 keys.

This is the calm before the storm. The next commit will switch from
Docopt to Clap.