raw: add get_key/get_key_into APIs
This provides the ability to map from a value back to its corresponding
key. This only works when values are monotonically increasing, which is
unfortunate, but it's still worth exposing. Since this is only useful is
limited circumstances, we don't make this available on the fst::Map
type, but instead provide it on raw::Fst only.

Closes #16