crc: add CRC32C to every FST
This modifies the FST builder to compute the CRC32C checksum as the FST
is being constructed. It is written as the last 4 bytes of the FST.

We also add a new `verify` routine on `raw::Fst` that permits callers to
check whether the FST's integrity is intact. Since verification could be
quite expensive for very large FSTs, we do not do this by default.

We avoid a dependency on a CRC32C crate since hand-rolling it ourselves
is very simple and not much code. I tried using a SIMD version of
CRC32C, but I couldn't benchmark a difference. In particular, I suspect
that the FST writing process is doing a lot of small writes, so the
checksummer doesn't get much opportunity to checksum a lot of bytes at
once.

It's not quite clear how to fix this. We could use our own buffer
internally, but then the caller wouldn't be able to use their own
buffered reader, which is a bit weird. But not without precedent.

In any case, the overhead of checksumming during construction is
virtually negligible, so we don't sweat it for now.