deps: remove byteorder
Now that std provides safe APIs for converting between raw bytes and
integers, we can use those instead. We still define a few utility
functions for convenience though.

The packing code is definitely not as ideally efficient as what
byteorder does, since we use safe code. But I was not able to observe a
difference in benchmarks.

fst now has no required dependencies!