Introduce a map_data method
This method is implemented on the fst::Map/Set and raw::Fst types, it allows
to change the underlying data type without having to deconstruct the type by
hand. This can be especially usefull when an fst Map or Set is based on another
type than a `Vec` and you need a default value for that type, you can call
default and map it to the right data type you need (e.g. `Arc`, `Rc`, `Cow`)
on a single line.