api: add map_data method
The new method is added to the fst::Map/Set and raw::Fst types. It
allows one to change the underlying data type without having to
deconstruct the type by hand. This can be especially usefull when an
fst Map or Set is based on a type other than a `Vec` and you need a
default value for that type, you can call default and map it to the
right data type you need (e.g. `Arc`, `Rc`, `Cow`) easily.

PR #108