Update node compile function
In the else branch there is a !node.is_final check.

However, in reviewing the code I found this confusing as it appears that node.is_final will always be false here. It will never be true at this point of the if branch.

If node.is_final is true it will always trigger the else if above.

So I removed the !node.is_final check.