impl: remove redundant check in 'compile'
In the else branch there is a !node.is_final check. However,
in the previous 'else if' block, there is a node.is_final check,
which guarantees that the 'else if' branch is taken whenever
'is_final' is true. Therefore, we know that 'is_final' is false in
the 'else' branch.

PR #128