ci: use 'cargo doc --all --no-deps'
'cargo doc' seems to now be warning on the fact that it's trying to
document both the "local" copy of 'fst' and also the copy of 'fst' from
crates.io. There was a bug in the Cargo.toml's of 'fst-bin' and 'bench'
in that they were specifying crates.io versions, so we remove that. But
it turns out that 'fst' from crates.io still makes it back into the
dependency tree via the 'csv' crate. (The 'csv' crate depends on
regex-automata, and since the levenshtein feature is enabled in one
place, it's enabled everywhere, and that pulls in 'fst'.)

Ref: https://github.com/rust-lang/cargo/issues/6313#issuecomment-855324163