Add a commandline option '--delimiter' to `map`.
This commandline option allows to specify the delimiter between keys and the values
in the input file to the `map` subcommand.

The provided delimiter should be a single byte utf-8 character.
If the value consists of multiple bytes only the first one is used.
This parameter is passed to `csv::ReaderBuilder::delimiter` which only
allows for a single byte as delimiter.