ci: bump pinned to Rust 1.60.0
Since fst-bin now relies on bstr 1.0, and bstr sets an MSRV of 1.60.0,
it looks like Cargo will refuse to select bstr 1.0 because of that.

The error message kind of sucks, and it is interesting that it prevents
building only 'fst' just because 'fst-bin' is in the same workspace. If
bstr didn't set 'rust-version' in its Cargo.toml, then 'cargo build' in
the fst repo root would still work (although building fst-bin would fail
because of the older Rust version innately).